requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
accelerate>=0.24.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
matplotlib>=3.7.0
pyyaml>=6.0.0
python-multipart>=0.0.6
lxml>=4.9.0
